+++ lib/lumberjack/device/writer.rb
@@ -55,9 +55,8 @@ module Lumberjack
       # :additional_lines and :time_format options will be passed through to the
       # Template constuctor.
       #
-      # The default template is "[:time :severity :progname(:pid) #:unit_of_work_id] :message"
-      # with additional lines formatted as "\n [#:unit_of_work_id] :message". The unit of
-      # work id will only appear if it is present.
+      # The default template is "[:time :severity :progname(:pid)] :message"
+      # with additional lines formatted as "\n  :message".
       #
       # The size of the internal buffer in bytes can be set by providing :buffer_size (defaults to 32K).
       #
@@ -68,12 +67,12 @@ module Lumberjack
         @stream = stream
         @stream.sync = true if @stream.respond_to?(:sync=)
         @buffer = Buffer.new
-        @buffer_size = (options[:buffer_size] || 0)
-        template = (options[:template] || DEFAULT_FIRST_LINE_TEMPLATE)
+        @buffer_size = options[:buffer_size] || 0
+        template = options[:template] || DEFAULT_FIRST_LINE_TEMPLATE
         if template.respond_to?(:call)
           @template = template
         else
-          additional_lines = (options[:additional_lines] || DEFAULT_ADDITIONAL_LINES_TEMPLATE)
+          additional_lines = options[:additional_lines] || DEFAULT_ADDITIONAL_LINES_TEMPLATE
           @template = Template.new(template, additional_lines: additional_lines, time_format: options[:time_format])
         end
       end
@@ -163,8 +162,6 @@ module Lumberjack
       def write_to_stream(lines)
         return if lines.empty?
         lines = lines.first if lines.is_a?(Array) && lines.size == 1
-
-        out = nil
         out = if lines.is_a?(Array)
           "#{lines.join(Lumberjack::LINE_SEPARATOR)}#{Lumberjack::LINE_SEPARATOR}"
         else