+++ spec/lumberjack/device/null_spec.rb
@@ -1,0 +1,10 @@
+require "spec_helper"
+
+RSpec.describe Lumberjack::Device::Null do
+  it "should not generate any output" do
+    device = Lumberjack::Device::Null.new
+    device.write(Lumberjack::LogEntry.new(Time.now, 1, "New log entry", nil, $$, nil))
+    device.flush
+    device.close
+  end
+end